Switch to azure active directory v1 end point
This resolves log in issues when users are logged in in more than one AD directory and / or on a microsoft service.

By default the v1 login page should ask users which account they want to sign in as. It is possible to provide a "domain hint" to the app to automatically select the correct account if the domain is known in advance. (REACT_APP_DOMAIN_HINT)