Add some notes about making it work locally
I used the current live setup in order to do this, but it still took quite a lot of fiddling. A reasonable amount of this fiddling is within the client\.env file, which has secrets in it and shouldn't go in to github. This is a bit of a shame. I might commit a .env.help file later with some notes