Remove npm installation from deployment
This was failing on azure. I ran the build commands on this commit and then  it to azure it works. This involves zipping everything up, with all the node_modules etc already done, and then uploading it to azure. Because I have already run all the install type scripts, azure doesn't need to do them on azure, which circumvents the problem. Still need to fix the problem though.