Remove check on label when sending emails
The label in projects.json is used by front and back end in different and conflicting ways.

It must be a label on the bug (eg 'user request') in order to send an email in response to a new comment

It must be the name of th repo in order to match a project on the front end (I think)

Neither of these is useful. I have removed it from the backend so now things can work. Removing it from the front end looks like a more involved job